server/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.2
sqlalchemy==2.0.23
pydantic==2.11.4
chromadb==0.4.22
sentence-transformers==3.0.1
# openai==1.3.7
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
numpy>=1.24.0,<2.0.0
pandas>=1.5.0 
boto3==1.39.17
langchain==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.1.1
langchain-postgres==0.0.15
psycopg2-binary==2.9.9
pgvector==0.2.5